requirements.txt
# ============================================
# Core LLM and Graph Framework
# ============================================
langchain>=0.3.7
langchain-community>=0.3.7
langchain-core>=0.3.17
langgraph>=0.2.45

# ============================================
# HuggingFace 경량 모델
# ============================================
transformers>=4.41.0
torch>=2.1.0
accelerate>=0.25.0
sentencepiece>=0.1.99
bitsandbytes>=0.41.0

# ============================================
# Word Document Generation
# ============================================
python-docx==1.1.2
lxml>=5.1.0

# ============================================
# Utilities
# ============================================
typing-extensions>=4.9.0
python-dotenv>=1.0.0
requests>=2.31.0

# ============================================
# FastAPI Web Framework 
# ============================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6